# docs/swagger.yaml
definitions:
  dal.AuthMessageTemplate:
    properties:
      channel:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      provider:
        type: string
      providerTemplateCode:
        type: string
      remark:
        type: string
      scene:
        type: string
      status:
        type: string
      subject:
        type: string
      tenantID:
        type: string
      updatedAt:
        type: string
    type: object
  dal.WxMpApp:
    properties:
      appID:
        type: string
      appSecret:
        type: string
      createdAt:
        type: string
      id:
        type: string
      remark:
        type: string
      status:
        type: string
      tenantID:
        type: string
      updatedAt:
        type: string
    type: object
  errcode.Error:
    properties:
      code:
        type: integer
      data: {}
      message:
        description: 用于存储自定义消息
        type: string
    type: object
...
  model.LoginReq:
    properties:
      email:
        description: 登录账号(输入邮箱或者手机号)
        example: test@test.cn
        type: string
      password:
        description: 密码
        example: "123456"
        maxLength: 512
        minLength: 6
        type: string
      salt:
        description: 随机盐(如果在超管设置了前端RSA加密则需要上送)
        maxLength: 512
        type: string
    required:
    - email
    - password
    type: object
  model.LoginRsp:
    properties:
      expires_in:
        description: 过期时间(单位:秒)
        type: integer
      token:
        description: 登录凭证
        type: string
    type: object
  model.Org:
    properties:
      address:
        description: 详细地址
        type: string
      city:
        description: 城市
        type: string
      contact_person:
        description: 联系人
        type: string
      created_at:
        description: 创建时间
        type: string
      district:
        description: 区县
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 组织ID
        type: string
      name:
        description: 组织名称
        type: string
      org_type:
        description: 组织类型
        type: string
      parent_id:
        description: 上级组织ID
        type: string
      phone:
        description: 电话
        type: string
      province:
        description: 省份
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态
        type: string
      tenant_id:
        description: 租户ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  model.OrgCreateReq:
    properties:
      address:
        description: 详细地址
        type: string
      city:
        description: 城市
        type: string
      contact_person:
        description: 联系人
        type: string
      district:
        description: 区县
        type: string
      email:
        description: 邮箱
        type: string
      name:
        description: 组织名称
        type: string
      org_type:
        description: '组织类型: BMS_FACTORY, PACK_FACTORY, DEALER, STORE'
        type: string
      parent_id:
        description: 上级组织ID
        type: string
      phone:
        description: 电话
        type: string
      province:
        description: 省份
        type: string
      remark:
        description: 备注
        type: string
    required:
    - name
    - org_type
    type: object
  model.OrgListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Org'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.OrgTreeNode:
    properties:
      address:
        description: 详细地址
        type: string
      children:
        items:
          $ref: '#/definitions/model.OrgTreeNode'
        type: array
      city:
        description: 城市
        type: string
      contact_person:
        description: 联系人
        type: string
      created_at:
        description: 创建时间
        type: string
      district:
        description: 区县
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 组织ID
        type: string
      name:
        description: 组织名称
        type: string
      org_type:
        description: 组织类型
        type: string
      parent_id:
        description: 上级组织ID
        type: string
      phone:
        description: 电话
        type: string
      province:
        description: 省份
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态
        type: string
      tenant_id:
        description: 租户ID
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  model.OrgUpdateReq:
    properties:
      address:
        description: 详细地址
        type: string
      city:
        description: 城市
        type: string
      contact_person:
        description: 联系人
        type: string
      district:
        description: 区县
        type: string
      email:
        description: 邮箱
        type: string
      name:
        description: 组织名称
        type: string
      phone:
        description: 电话
        type: string
      province:
        description: 省份
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: '状态: N-正常, F-禁用'
        type: string
    type: object
  model.Response:
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 业务数据
      message:
        description: 提示信息
        example: success
        type: string
    type: object
  model.UpdateUserAddressReq:
    properties:
      additional_info:
        description: 附加信息
        maxLength: 500
        type: string
      address_label:
        description: 地址标签
        maxLength: 50
        type: string
      city:
        description: 城市
        maxLength: 50
        type: string
      country:
        description: 国家
        maxLength: 50
        type: string
      detailed_address:
        description: 详细地址
        maxLength: 200
        type: string
      district:
        description: 区县
        maxLength: 50
        type: string
      latitude:
        description: 纬度
        maxLength: 20
        type: string
      longitude:
        description: 经度
        maxLength: 20
        type: string
      postal_code:
        description: 邮政编码
        maxLength: 10
        type: string
      province:
        description: 省份
        maxLength: 50
        type: string
      street:
        description: 街道/乡镇
        maxLength: 100
        type: string
    type: object
  model.UpsertAuthMessageTemplateReq:
    properties:
      channel:
        enum:
        - EMAIL
        - SMS
        maxLength: 16
        type: string
      content:
        maxLength: 10000
        type: string
      provider:
        maxLength: 36
        type: string
      provider_template_code:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      scene:
        enum:
        - LOGIN
        - REGISTER
        - RESET_PASSWORD
        - BIND
        maxLength: 32
        type: string
      status:
        enum:
        - OPEN
        - CLOSE
        maxLength: 16
        type: string
      subject:
        maxLength: 200
        type: string
    required:
    - channel
    - scene
    - status
    type: object
  model.UpsertWxMpAppReq:
    properties:
      app_secret:
        maxLength: 128
        type: string
      appid:
        maxLength: 64
        type: string
      remark:
        maxLength: 255
        type: string
      status:
        enum:
        - OPEN
        - CLOSE
        maxLength: 16
        type: string
    required:
    - app_secret
    - appid
    - status
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: ThingsPanel API.
  title: ThingsPanel API
  version: "1.0"
paths:
...
  /api/v1/app/auth/config/templates:
    get:
      parameters:
      - description: SYS_ADMIN指定租户
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dal.AuthMessageTemplate'
            type: array
      tags:
      - APP-Auth-Config
    post:
      consumes:
      - application/json
      parameters:
      - description: SYS_ADMIN指定租户
        in: query
        name: tenant_id
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpsertAuthMessageTemplateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      tags:
      - APP-Auth-Config
  /api/v1/app/auth/config/wxmp:
    get:
      parameters:
      - description: SYS_ADMIN指定租户
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dal.WxMpApp'
      tags:
      - APP-Auth-Config
    post:
      consumes:
      - application/json
      parameters:
      - description: SYS_ADMIN指定租户
        in: query
        name: tenant_id
        type: string
      - description: 请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpsertWxMpAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      tags:
      - APP-Auth-Config
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 认证令牌(Token)将在用户成功登录后生成并返回。客户端需要在后续所有需要认证的API请求中，将此令牌添加到HTTP请求头(Header)的'x-token'字段中。服务器将通过验证此令牌来确认用户身份并授权访问受保护资源。
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.LoginRsp'
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户登录
      tags:
      - 用户认证
...
  /api/v1/org:
    get:
      consumes:
      - application/json
      description: 分页查询组织列表，支持按类型筛选
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: '组织类型: BMS_FACTORY, PACK_FACTORY, DEALER, STORE'
        in: query
        name: org_type
        type: string
      - description: 组织名称
        in: query
        name: name
        type: string
      - description: '状态: N-正常, F-禁用'
        in: query
        name: status
        type: string
      - description: 父组织ID
        in: query
        name: parent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.OrgListResp'
      summary: 获取组织列表
      tags:
      - 组织管理
    post:
      consumes:
      - application/json
      description: 创建新的组织（BMS厂家/PACK厂家/经销商/门店）
      parameters:
      - description: 组织信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.OrgCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Org'
      summary: 创建组织
      tags:
      - 组织管理
  /api/v1/org/{id}:
    delete:
      consumes:
      - application/json
      description: 删除组织（需确保无子组织和关联设备）
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除组织
      tags:
      - 组织管理
    get:
      consumes:
      - application/json
      description: 根据ID获取组织详细信息
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Org'
      summary: 获取组织详情
      tags:
      - 组织管理
    put:
      consumes:
      - application/json
      description: 更新组织信息（不支持修改上级组织）
      parameters:
      - description: 组织ID
        in: path
        name: id
        required: true
        type: string
      - description: 组织信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.OrgUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新组织
      tags:
      - 组织管理
  /api/v1/org/tree:
    get:
      consumes:
      - application/json
      description: 获取组织树结构，可按类型筛选
      parameters:
      - description: 组织类型筛选
        in: query
        name: org_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.OrgTreeNode'
            type: array
      summary: 获取组织树
      tags:
      - 组织管理
...
  /api/v1/user/address/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定用户的地址信息，支持创建新地址或更新现有地址
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 地址信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserAddressReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: object
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新用户地址信息
      tags:
      - 用户管理
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"